* 🎨 style(baseof.html): adjust spacing and add Github icon The spacing between the footer links has been adjusted to improve readability. The RSS link has been moved to the right of the Blog link and is now only visible on medium-sized screens and above. A Github icon has been added to the Github link to improve visual consistency with other social media links.
* 🎨 style(baseof.html): replace social media icons with SVG icons
The social media icons have been replaced with SVG icons. This improves the accessibility of the website as SVG icons are more accessible than image icons.

* 🎨 style(baseof.html): replace text with SVG icon for Layers link
The Layers link in the header has been updated to use an SVG icon instead of text. This improves the visual design of the header and makes it more consistent with the rest of the site.

* 🎨 style(index.html): add height to design link container
The design link container now has a height of 40px, which improves the visual consistency of the page.